Uploading blargboard (master)

+++ pages/forum.php
@@ -1,6 +1,7 @@
 <?php
 //  AcmlmBoard XD - Thread listing page
 //  Access: all
+if (!defined('BLARG')) die();
 
 if(!isset($_GET['id']))
 	Kill(__("Forum ID unspecified."));
@@ -10,52 +11,57 @@ $fid = (int)$_GET['id'];
 if (!HasPermission('forum.viewforum', $fid))
 	Kill(__('You may not access this forum.'));
 
-if($loguserid && $_GET['action'] == "markasread")
-{
-	Query("REPLACE INTO {threadsread} (id,thread,date) SELECT {0}, {threads}.id, {1} FROM {threads} WHERE {threads}.forum={2}",
-		$loguserid, time(), $fid);
-
-	die(header("Location: ".actionLink("board")));
-}
-
-
 $rFora = Query("select * from {forums} where id={0}", $fid);
 if(NumRows($rFora))
 	$forum = Fetch($rFora);
 else
 	Kill(__("Unknown forum ID."));
+	
+if ($forum['redirect'])
+	die(header('Location: '.forumRedirectURL($forum['redirect'])));
+	
+if($loguserid)
+{
+	if($_GET['action'] == "markasread")
+	{
+		Query("REPLACE INTO {threadsread} (id,thread,date) SELECT {0}, {threads}.id, {1} FROM {threads} WHERE {threads}.forum={2}",
+			$loguserid, time(), $fid);
+
+		die(header("Location: ".actionLink("board", $forum['board'])));
+	}
+	
+	$isIgnored = FetchResult("select count(*) from {ignoredforums} where uid={0} and fid={1}", $loguserid, $fid) == 1;
+	if(isset($_GET['ignore']))
+	{
+		if(!$isIgnored)
+			Query("insert into {ignoredforums} values ({0}, {1})", $loguserid, $fid);
+		die(header("Location: ".$_SERVER['HTTP_REFERER']));
+	}
+	else if(isset($_GET['unignore']))
+	{
+		if($isIgnored)
+			Query("delete from {ignoredforums} where uid={0} and fid={1}", $loguserid, $fid);
+		die(header("Location: ".$_SERVER['HTTP_REFERER']));
+	}
+}
 
 $title = $forum['title'];
 $urlname = HasPermission('forum.viewforum', $fid, true) ? $title : '';
 
 
-$isIgnored = FetchResult("select count(*) from {ignoredforums} where uid={0} and fid={1}", $loguserid, $fid) == 1;
-if(isset($_GET['ignore']))
-{
-	if(!$isIgnored)
-		Query("insert into {ignoredforums} values ({0}, {1})", $loguserid, $fid);
-	die(header("Location: ".actionLink("forum", $fid)));
-}
-else if(isset($_GET['unignore']))
-{
-	if($isIgnored)
-		Query("delete from {ignoredforums} where uid={0} and fid={1}", $loguserid, $fid);
-	die(header("Location: ".actionLink("forum", $fid)));
-}
-
-$links = '';
+$links = array();
 if($loguserid)
-	$links .= actionLinkTagItem(__("Mark forum read"), "forum", $fid, "action=markasread", $urlname);
+	$links[] = actionLinkTag(__("Mark forum read"), "forum", $fid, "action=markasread", $urlname);
 
 if($loguserid)
 {
 	if($isIgnored)
-		$links .= "<li>".actionLinkTag(__("Unignore forum"), "forum", $fid, "unignore", $urlname)."</li>";
+		$links[] = actionLinkTag(__("Unignore forum"), "forum", $fid, "unignore", $urlname);
 	else
-		$links .= "<li>".actionLinkTag(__("Ignore forum"), "forum", $fid, "ignore", $urlname)."</li>";
+		$links[] = actionLinkTag(__("Ignore forum"), "forum", $fid, "ignore", $urlname);
 
 	if (HasPermission('forum.postthreads', $fid))
-		$links .= "<li>".actionLinkTag(__("Post thread"), "newthread", $fid, '', $urlname)."</li>";
+		$links[] = actionLinkTag(__("Post thread"), "newthread", $fid, '', $urlname);
 }
 
 $metaStuff['description'] = htmlspecialchars(strip_tags($forum['description']));
@@ -64,7 +70,10 @@ $metaStuff['tags'] = getKeywords(strip_tags($forum['title']));
 $OnlineUsersFid = $fid;
 MakeCrumbs(forumCrumbs($forum), $links);
 
+makeAnncBar();
+
 makeForumListing($fid);
+
 
 $total = $forum['numthreads'];
 $tpp = $loguser['threadsperpage'];
@@ -88,49 +97,16 @@ $rThreads = Query("	SELECT
 					WHERE forum={0}
 					ORDER BY sticky DESC, lastpostdate DESC LIMIT {1u}, {2u}", $fid, $from, $tpp, $loguserid);
 
-$numonpage = NumRows($rThreads);
-
 $pagelinks = PageLinks(actionLink("forum", $fid, "from=", $urlname), $tpp, $from, $total);
-
-if($pagelinks)
-	echo "<div class=\"smallFonts pages\">".__("Pages:")." ".$pagelinks."</div>";
 	
-makeAnncBar();
-
 $ppp = $loguser['postsperpage'];
 if(!$ppp) $ppp = 20;
 
 if(NumRows($rThreads))
 {
-	$forumList = "";
-	$haveStickies = 0;
-	$cellClass = 0;
-
-	while($thread = Fetch($rThreads))
-	{
-		$forumList .= listThread($thread, $cellClass);
-		$cellClass = ($cellClass + 1) % 2;
-	}
-
-	Write(
-"
-	<table class=\"outline margin width100\">
-		<tr class=\"header1\">
-			".($mobileLayout?
-			"<th>".__('Thread')."</th>":
-			"<th style=\"width: 28px;\">&nbsp;</th>
-			<th style=\"width: 16px;\">&nbsp;</th>
-			<th style=\"width: 60%;\">".__("Title")."</th>
-			<th>".__("Started by")."</th>
-			<th>".__("Replies")."</th>
-			<th>".__("Views")."</th>
-			<th style=\"min-width:150px\">".__("Last post")."</th>"
-			)."
-		</tr>
-		{0}
-	</table>
-",	$forumList);
-} else
+	makeThreadListing($rThreads, $pagelinks);
+} 
+else
 	if(!HasPermission('forum.postthreads', $fid))
 		Alert(__("You cannot start any threads here."), __("Empty forum"));
 	elseif($loguserid)
@@ -138,13 +114,7 @@ if(NumRows($rThreads))
 	else
 		Alert(format(__("{0} so you can post something."), actionLinkTag(__("Log in"), "login")), __("Empty forum"));
 
-if($pagelinks)
-	Write("<div class=\"smallFonts pages\">".__("Pages:")." {0}</div>", $pagelinks);
-
-if (!$mobileLayout)
-{
-	ForumJump();
-}
+ForumJump();
 
 
 function fj_forumBlock($fora, $catid, $selID, $indent)
@@ -154,16 +124,7 @@ function fj_forumBlock($fora, $catid, $selID, $indent)
 	foreach ($fora[$catid] as $forum)
 	{
 		if ($forum['redirect'])
-		{
-			$redir = $forum['redirect'];
-			if ($redir[0] == ':')
-			{
-				$redir = explode(':', $redir);
-				$forumlink = actionLink($redir[1], $redir[2], $redir[3], $redir[4]);
-			}
-			else
-				$forumlink = $redir;
-		}
+			$forumlink = forumRedirectURL($forum['redirect']);
 		else
 			$forumlink = actionLink('forum', $forum['id'], '', HasPermission('forum.viewforum',$forum['id'],true)?$forum['title']:'');
 			
@@ -215,17 +176,13 @@ function ForumJump()
 '			</optgroup>
 ';
 	}
+	
+	$theList = '<select onchange="document.location=this.options[this.selectedIndex].value;">'
+		.($forum['board']?'<option value="'.actionLink('board').'">Back to main forums</option>':'')
+		.$theList
+		.'</select>';
 
-	write(
-"
-	<label>
-		".__("Forum Jump:")."
-		<select onchange=\"document.location=this.options[this.selectedIndex].value;\">
-			{1}
-			{0}
-		</select>
-	</label>
-",	$theList, ($forumpage != 0) ? '<option value="'.actionLink('board').'">Back to main forums</option>':'');
+	RenderTemplate('forumjump', array('forumlist' => $theList));
 }
 
 ?>
